+++ guillotina/tests/test_metrics.py
@@ -1,7 +1,11 @@
 from asyncmock import AsyncMock
+from guillotina import metrics
 from guillotina.contrib.redis.driver import RedisDriver
 from guillotina.db.storages.pg import PostgresqlStorage
 from unittest.mock import MagicMock
+
+import asyncio
+import prometheus_client
 
 
 class TestRedisMetrics:
@@ -78,9 +82,14 @@ class TestRedisMetrics:
 
 
 class TestPGMetrics:
+    def _make_txn(self):
+        mock = AsyncMock()
+        mock.lock = mock._lock = asyncio.Lock()
+        return mock
+
     async def test_load_object(self, metrics_registry):
         storage = PostgresqlStorage()
-        await storage.load(AsyncMock(), "foobar")
+        await storage.load(self._make_txn(), "foobar")
         assert (
             metrics_registry.get_sample_value(
                 "guillotina_db_pg_ops_total", {"type": "load_object_by_oid", "error": "none"}
@@ -93,13 +102,19 @@ class TestPGMetrics:
             )
             > 0
         )
+        assert (
+            metrics_registry.get_sample_value(
+                "guillotina_db_pg_lock_time_seconds_sum", {"type": "load_object_by_oid"}
+            )
+            > 0
+        )
 
     async def test_store_object(self, metrics_registry):
         storage = PostgresqlStorage(store_json=False)
         ob = MagicMock()
         ob.__new_marker__ = False
         ob.__serial__ = 1
-        txn = AsyncMock()
+        txn = self._make_txn()
         txn.get_connection.return_value.fetch.return_value = [{"count": 1}]
         await storage.store("foobar", 1, MagicMock(), ob, txn)
         assert (
@@ -114,10 +129,22 @@ class TestPGMetrics:
             )
             > 0
         )
+        assert (
+            metrics_registry.get_sample_value(
+                "guillotina_db_pg_lock_time_seconds_sum", {"type": "store_object"}
+            )
+            > 0
+        )
+        assert (
+            metrics_registry.get_sample_value(
+                "guillotina_db_pg_lock_time_seconds_sum", {"type": "store_object"}
+            )
+            > 0
+        )
 
     async def test_delete_object(self, metrics_registry):
         storage = PostgresqlStorage(autovacuum=False)
-        await storage.delete(AsyncMock(), "foobar")
+        await storage.delete(self._make_txn(), "foobar")
         assert (
             metrics_registry.get_sample_value(
                 "guillotina_db_pg_ops_total", {"type": "delete_object", "error": "none"}
@@ -130,3 +157,26 @@ class TestPGMetrics:
             )
             > 0
         )
+        assert (
+            metrics_registry.get_sample_value(
+                "guillotina_db_pg_lock_time_seconds_sum", {"type": "delete_object"}
+            )
+            > 0
+        )
+        assert (
+            metrics_registry.get_sample_value(
+                "guillotina_db_pg_lock_time_seconds_sum", {"type": "delete_object"}
+            )
+            > 0
+        )
+
+
+class TestLock:
+    async def test_lock_metric(self):
+        lock = asyncio.Lock()
+        metric = prometheus_client.Histogram("test_metric", "Test",)
+        assert metric.collect()[0].samples[0].value == 0
+        async with metrics.watch_lock(metric, lock):
+            assert lock.locked()
+        assert not lock.locked()
+        assert metric.collect()[0].samples[0].value == 1