+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ CHANGELOG
 5.3.51 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Record metrics on time waiting for pg locks
+  [vangheem]
 
 
 5.3.50 (2020-09-23)